ref(chat): Simplify sandbox session flow
Split sandbox lifecycle, skill sync, eval shim, and error classification into focused sandbox modules while keeping the executor surface unchanged.

Flatten the session manager control flow so sandbox acquisition and bash execution read top-down instead of mixing state resets, retry branches, and temporary policy handling in one block.

Preserve existing sandbox lifecycle behavior while making the remaining complexity live in smaller, easier-to-read files.
